argument dimid in ncmpii_elem_NC_dimarray() should be of type int
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@1725 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/lib/nc.h
@@ -161,7 +161,7 @@ extern int
 ncmpii_dup_NC_dimarray(NC_dimarray *ncap, const NC_dimarray *ref);
 
 extern NC_dim *
-ncmpii_elem_NC_dimarray(const NC_dimarray *ncap, size_t elem);
+ncmpii_elem_NC_dimarray(const NC_dimarray *ncap, int elem);
 
 extern int
 ncmpi_def_dim(int ncid, const char *name, MPI_Offset size, int *dimidp);
